+++ src/components/Header/SubHeader.vue
@@ -177,6 +177,7 @@ export default class SubHeader extends DaintreeComponent {
       { name: "Elastic IPs", link: "#/network/eips" },
       { name: "Security Groups", link: "#/network/securityGroups" },
       { name: "Peering connections", link: "#/network/peeringConnections" },
+      { name: "Endpoints", link: "#/network/endpoints" },
     ],
     EC2: [
       { name: "Instances", link: "#/ec2/instances" },